requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
Django==2.2.3
django-handlers==0.1.1
django-hitcount==1.3.1
django-sendgrid-v5==0.8.0
future==0.17.1
importlib-metadata==0.19
kombu==4.6.4
psycopg2==2.8.3
pytz==2019.1
redis==3.3.8
sendgrid==6.0.5
django-sendgrid-v5==0.8.0
django-choices==1.7.0
sqlparse==0.3.0
vine==1.3.0
zipp==0.5.2
django-choices==1.7.0

django-sendgrid-v5==0.8.0
django-hitcount==1.3.1
